requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-WTF>=1.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
kiteconnect>=4.0.0
Werkzeug>=3.0.0
WTForms>=3.0.0
email-validator>=2.0.0
openpyxl>=3.0.0
pandas>=2.0.0
